pyproject.toml
[project]
name = "exec-radar"
version = "0.1.0"
description = "AI-powered executive opportunity intelligence platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "httpx>=0.27,<1",
    "pyyaml>=6,<7",
    "jinja2>=3.1,<4",
    "sqlalchemy[asyncio]>=2,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5,<6",
    "pytest-playwright>=0.6,<1",
    "ruff>=0.8,<1",
    "mypy>=1.13,<2",
    "pre-commit>=4,<5",
    "aiosqlite>=0.21,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "packages"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["apps", "packages", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "T20", # flake8-print
    "RUF", # ruff-specific rules
]
ignore = ["S101"]  # allow assert in tests

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S106"]  # board_token args in tests are public API identifiers, not passwords

[tool.ruff.lint.isort]
known-first-party = ["apps", "packages"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
# Playwright e2e tests use a session-scoped event loop that conflicts with
# pytest-asyncio's function-scoped loops. Run them separately:
#   pytest tests/e2e/ -v
addopts = "--ignore=tests/e2e"

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
